+++ js/field-action.js
@@ -1,0 +1,1 @@
+import './map.js';
