+++ js/form-action.js
@@ -1,0 +1,1 @@
+import './field-action.js';
