+++ js/page-state.js
@@ -1,0 +1,2 @@
+import './form-action.js';
+import './map.js';
